Flytter på test av oppgave endret til egen fil og litt refaktorering
